requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
djangorestframework-word-filter==0.2.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
